+++ test/unit/fuzz-tools.test.js
@@ -334,6 +334,7 @@ import * as tonnisHipOaV354 from '../../lib/tonnis-hip-oa-v354.js';
 import * as lachmanAclV355 from '../../lib/lachman-acl-v355.js';
 import * as ceapVenousV356 from '../../lib/ceap-venous-v356.js';
 import * as nyhaClassV357 from '../../lib/nyha-class-v357.js';
+import * as ramsaySedationV358 from '../../lib/ramsay-sedation-v358.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -613,6 +614,7 @@ const MODULES = {
   'lachman-acl-v355.js': lachmanAclV355,
   'ceap-venous-v356.js': ceapVenousV356,
   'nyha-class-v357.js': nyhaClassV357,
+  'ramsay-sedation-v358.js': ramsaySedationV358,
   'tb-testing.js': tbTesting,
 };